backend/requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
pydantic==2.12.5
opencv-python-headless==4.11.0.86
deepface==0.0.98
tensorflow==2.15.0
tf-keras==2.15.0
numpy==1.26.4
mediapipe==0.10.9
pydub
onnxruntime
silero-vad
torch
python-multipart==0.0.22
python-socketio
anyio
